Renamed trail.h to eulertrail.h
