+++ include/sealib/eulertrail.h
